Implement bind operation
The simplest way to implement authentication for a bucket without giving
the broker the ability to create IAM policies is to name a specific user
in a bucket policy. Therefore, we implement binding by creating an IAM
user and keys and putting a bucket policy which names them as the
`Principal`.

We have to put a retry mechanism into the putting of the policy because
IAM and S3 take a number of seconds to get in-sync with each other.

Also, we have to do some awkward type juggling because the policies are
returned via the Golang SDK as `*string` types, and the AWS Principal
may be a string or an array of strings, depending on whether there are
one or more existing principals.

Currently the S3 client integration tests leave an orphaned IAM user.
This should be fixed in an upcoming commit.

+++ testing/integration/s3client/s3client_test.go
@@ -1,22 +1,40 @@
 package s3client_test
 
 import (
+	"io/ioutil"
+	"os"
+	"path/filepath"
+
 	"github.com/alphagov/paas-s3-broker/s3"
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/credentials"
+	"github.com/aws/aws-sdk-go/aws/session"
+	awsS3 "github.com/aws/aws-sdk-go/service/s3"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-
 	"github.com/satori/go.uuid"
+)
+
+const (
+	region       = "eu-west-2"
+	bucketPrefix = "gds-paas-s3-broker-testing-"
+	fixturesPath = "../../fixtures/"
+	testFileKey  = "file.txt"
 )
 
 var _ = Describe("S3client", func() {
 	var (
 		s3Client   *s3.S3Client
 		bucketName string
+		user1      string
+		user2      string
 	)
 
 	BeforeEach(func() {
-		s3Client = s3.NewS3Client("gds-paas-s3-broker-testing-", "eu-west-2")
+		s3Client = s3.NewS3Client(bucketPrefix, region)
 		bucketName = uuid.NewV4().String()
+		user1 = uuid.NewV4().String()
+		user2 = uuid.NewV4().String()
 	})
 
 	It("can manage the S3 bucket lifecycle", func() {
@@ -24,8 +42,60 @@ var _ = Describe("S3client", func() {
 		err := s3Client.CreateBucket(bucketName)
 		Expect(err).NotTo(HaveOccurred())
 
+		By("adding a user to the bucket policy")
+		user1Creds, err := s3Client.AddUserToBucket(user1, bucketName)
+		Expect(err).NotTo(HaveOccurred())
+
+		By("using the returned credentials to prove access")
+		assertBucketAccess(user1Creds, bucketName)
+
+		By("adding another user to the bucket policy")
+		user2Creds, err := s3Client.AddUserToBucket(user2, bucketName)
+		Expect(err).NotTo(HaveOccurred())
+
+		By("using the returned credentials to prove access")
+		assertBucketAccess(user2Creds, bucketName)
+
+		By("preserving the first user's access")
+		assertBucketAccess(user1Creds, bucketName)
+
+		By("deleting the user")
+
 		By("deleting the bucket")
 		err = s3Client.DeleteBucket(bucketName)
 		Expect(err).NotTo(HaveOccurred())
 	})
 })
+
+func assertBucketAccess(creds s3.BucketCredentials, bucketName string) {
+	sess := session.Must(session.NewSession(&aws.Config{
+		Region:      aws.String(region),
+		Credentials: credentials.NewStaticCredentials(creds.AWSAccessKeyID, creds.AWSSecretAccessKey, ""),
+	}))
+	userS3Client := awsS3.New(sess)
+	file, err := os.Open(filepath.Join(fixturesPath, testFileKey))
+	Expect(err).NotTo(HaveOccurred())
+	defer file.Close()
+
+	_, err = userS3Client.PutObject(&awsS3.PutObjectInput{
+		Bucket: aws.String(bucketPrefix + bucketName),
+		Key:    aws.String(testFileKey),
+		Body:   file,
+	})
+	Expect(err).NotTo(HaveOccurred())
+
+	getObjectOutput, err := userS3Client.GetObject(&awsS3.GetObjectInput{
+		Bucket: aws.String(bucketPrefix + bucketName),
+		Key:    aws.String(testFileKey),
+	})
+	Expect(err).NotTo(HaveOccurred())
+	fileContents, err := ioutil.ReadAll(getObjectOutput.Body)
+	Expect(err).NotTo(HaveOccurred())
+	Expect(string(fileContents)).To(ContainSubstring("This is a file"))
+
+	_, err = userS3Client.DeleteObject(&awsS3.DeleteObjectInput{
+		Bucket: aws.String(bucketPrefix + bucketName),
+		Key:    aws.String(testFileKey),
+	})
+	Expect(err).NotTo(HaveOccurred())
+}
